Filter out unavailable LPSolverLibs from solvers list
oscar-linprog tests assume all the external solvers are available
but this is not necessarily the case (especially for the commercial Gurobi).
OscaR supplies the relevant jars need to compile the bindings and test-cases but
the test cases should only be executed against the solvers you have configured.

Created a helper method that attempts to instantiate an LPSolver instance. If it
encounters an UnsatisfiedLinkError (e.g. you don?t have the native binary) then
that option is removed from the solvers used in the test cases.