Added support for creating modulo constraints using the convenient syntax: x % y == z
Achieved this by overriding % operate of CPIntVar to create a case class representing
the left hand side of the modulo expression. This class (ModuloLHS) defines == operators
for CPIntVar and Int that create a Modulo constraint with the appropriate values.

Amended the test cases in TestModulo to use the new syntax and it all seems to work.